[libdrm] fix random system hang in decode/encode
mos_bo_get_datatype does not report correct feature if i915 driver does not support it. This leaves the feature randomly on/off and causes random system hang.
This change fixes the issue by clearing the feature when i915 driver does not support it. Further fix will be in media driver which calls it.

Change-Id: Ibc8b262d6219fdece07fecd597eed5c994c57b5d