[CM/SIM] Implements dual context support.(CL#744769)
Users can create a CmQueue via either CmDevice::CreateQueue() or CmDevice::CreateQueueEx().
For CmDevice::CreateQueue(), a CmQueue corresponding to the render context is returned by default. In order to use the compute context, please set the environment variable "CM_USE_COMPUTE_CONTEXT" to a positive value. This function then returns a CmQueue with the compute context.
CmDevice::CreateQueueEx() creates a CmQueue according to the option. Users need to explicitly specify the type in parameters.

Change-Id: If614ce0bbb3894ff81c917fade7f217abb576f34