Add ENABLE_KERNELS build option cmake -DENABLE_KERNELS=OFF <workspace>/media-driver
This build option permits to exclude all shaders (kernels) from
the driver build. This provides a full open source driver build.

If someone wants to make sure that binary shaders don’t present
for sure it is possible to delete them with:
  # find . -name igcodeckrn*.c | xargs git rm
  # find . -name igvpkrn*.c | xargs git rm
Mind please that it is not needed to remove corresponding .h files
since they will be needed for successful compilation. The trick is that
0-dummy shaders will actually be fetched from
media_driver/agnostic/common/shared/dummykrn.c
By the build system

Change-Id: I90d6ea7d9dc4df01f6a3741daedc97ad422d6467
Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>