[CM] Force the hybrid read and write to be blocking operations
Two events were needed in hybrid read and write, one for sync before
copy and the other to wait till the copy ends. As we only have one event
in API, we will use this API as to sync before copy, as normal
read/write does, and don't need the second event as we force it to be
blocking operation.

Change-Id: Ibc2a63f8d9a98c47a9ebc347ef41dd5d10515c7f