[build] Keep media driver versioning in sources
Fixes: #834
This change keeps media driver version in sources in a top
level cmake file. This version is not modifiable. If user
needs to customes the version, he can specify the following
variable in a cmake command line:
  cmake -DMEDIA_VERSION_EXTRA=-patched
If media driver version in sources is 20.1.pre2 the above
user adjustment will result in 20.1.pre2-patched version
string. This version string gets reported by the driver and
can be seen in vainfo output:
  Intel iHD driver for Intel(R) Gen Graphics - 20.1.pre2 (abcdefg)
Where abcdefg is git commit id.

Change-Id: Ib2c66210cab08df1b3654b429dcffa9574da19e8